configuração de socket.io no projeto web.

+++ web/src/components/DevItem/index.js
@@ -1,19 +1,53 @@
-import React from 'react';
-import { MdClose } from 'react-icons/md';
+import React, { useState, useEffect } from 'react';
+import { MdClose, MdEdit } from 'react-icons/md';
+import { FaGithub } from 'react-icons/fa';
+import Swal from 'sweetalert2';
 
 import './styles.css';
 
-export default function DevItem({ dev, onDeleteForm }) {
+export default function DevItem({ dev, onDeleteForm, onEditForm }) {
+  const [edit, setEdit] = useState(false);
+
+  useEffect(() => {
+    async function editForm() {
+      await onEditForm(edit);
+    }
+
+    editForm();
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [edit]);
+
   const { avatar_url, name, techs, github_username, bio } = dev;
 
   async function handleRemoveUser() {
-    await onDeleteForm({ github_username });
+    await Swal.fire({
+      title: 'Remover usuário',
+      text: 'Você tem certeza que quer remover o usuário?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      reverseButtons: true,
+      confirmButtonText: 'Sim, remover usuário!',
+    }).then(async result => {
+      if (result.value) {
+        await onDeleteForm({ github_username });
+        Swal.fire('Removido!', 'Usuário removido com sucesso.', 'success');
+      }
+    });
+  }
+
+  async function handleEditUser() {
+    await setEdit(!edit);
   }
 
   return (
     <li className="dev-item">
       <div className="remove-item">
-        <button type="button" onClick={() => handleRemoveUser()}>
+        <button
+          title="Deletar"
+          type="button"
+          onClick={() => handleRemoveUser()}
+        >
           <MdClose />
         </button>
       </div>
@@ -26,9 +60,17 @@ export default function DevItem({ dev, onDeleteForm }) {
       </header>
 
       <p>{bio}</p>
-      <a href={`https://github.com/${github_username}`}>
-        Acessar perfil no Github
-      </a>
+      <div className="plus">
+        <div className="github-link">
+          <FaGithub />
+          <a href={`https://github.com/${github_username}`}>
+            Acessar perfil no Github
+          </a>
+        </div>
+        <button title="Editar" type="button" onClick={() => handleEditUser()}>
+          <MdEdit />
+        </button>
+      </div>
     </li>
   );
 }
